"Various modifications to the basic Elo system, mostly heuristic in nature, have been proposed to deal with the issues of draws, team games, and multiple players. A draw, for example, can be represented as a score of 1/2 in computing the skill updates. The fundamental problem is that, since Elo is an algorithm rather than a model, it is difficult to see what assumptions are being made by making changes to the algorithm. Without understanding the assumptions underlying the each change it is hard to predict how well the resulting algorithm will work – and, more importantly, extremely hard to diagnose problems with the algorithm when they arise. A model-based approach allows such extensions to be incorporated in a transparent way, giving rise to a solution which can handle all of the above complexities – whilst remaining both understandable and maintainable."
This book really makes a good story for why model based ML is necessary. I am guessing same applies to model based RL.

When I am done with the book, I am going to do some research on pruning.

I think pruning is the breaking point. I want to understand what could be a possible candidate for which synapses get pruned because I cannot even imagine how it could be done. If there aren't any ways of doing it, then the backprop algorithm is wrong as a basis. It is probably just a small part of a greater whole.