"5:10pm. "You may think that our online learning process updates our skill distribution for a player over time and so would allow the skill to change. This is a common misconception about online learning, but it is not true. Our current model assumes that the skill of a player is a fixed, but unknown, quantity. Online learning does not represent the modelling of an evolving skill value, but rather an updating of the uncertainty in this unknown fixed-across-time skill."
I wonder if this is why the squared error was giving me erroneous results when estimating the outputs?

5:20pm. "This particular model, in which we assume a Gaussian-distributed change is added at each stage is known as the Kalman filter [Kalman, 1960; Zarchan and Musoff, 2005] and is widely used in signal processing. Note that we do not need to know anything about Kalman filters, or to be familiar with the literature or with the associated algorithms or terminology, in order to use them in our application. We just write down the model corresponding to our assumptions and run online message passing, and this will implement the standard Kalman filtering equations automatically."

Kalman filters are this simple? They certainly did not seem that way when I last saw them.

5:25pm. "Typically this is done by selecting players for which the game outcome is most uncertain. Note that this also tends to produce matches whose outcomes are the most informative in terms of learning the skills of the players."

I read something similar to this in the context of coevolution. I forgot what the name of the guy giving the talk was. He had a great quote about forgetting being the cause of cycles in dynamical systems.

https://papers.nips.cc/paper/3331-trueskill-through-time-revisiting-the-history-of-chess
TrueSkill Through Time - Revisiting the History of Chess

I'll read this paper later.

...I took a peek.

5:55pm. http://mbmlbook.com/EmailClassifier_A_model_for_classification.html

This could be amazing. So he is building a NN here for spam classification, except from a Bayesian perspective. Yeah, I need this stuff.

"Unfortunately, if we attempt to run inference on this model then any inference algorithm we try will fail. This is because some of the observed values have zero probability under the model. In other words, there is no way that the data-generating process encoded by our model could have generated the observed data values. When your data has zero probability under your model, it is a sure sign that the model is wrong!"

7pm. https://arxiv.org/abs/1511.03405
Training Deep Gaussian Processes using Stochastic Expectation Propagation and Probabilistic Backpropagation

https://arxiv.org/abs/1502.05336
Probabilistic Backpropagation for Scalable Learning of Bayesian Neural Networks

https://faculty.ozyegin.edu.tr/melihkandemir/files/2017/05/sml_lecture10.pdf
Statistical Machine Learning
Lecture 10: Expectation Propagation and
Probabilistic Backpropagation

I'll take a look at this tomorrow. Let me finish the chapter and I will call it a day.

I like how in factor graphs the weights have a distribution rather than point. That might have some potential.

Wolfgang Maas did claim that weights in brains are sampled from a distribution.

But the real thing that I need to figure out is the Bayesian perspective on RL. Just what the hell am I even doing in RL?

https://faculty.ozyegin.edu.tr/melihkandemir/sml-course/

I'll take a look at the pdfs for this later.

To be honest, this is very hard stuff, but I've been doing this for over 3.5 years now. Somehow I feel that I might be able to digest this, if I pick the right trajectory.

7:40pm. This stuff in figure 4.26 is quite something. I still need to wrap my this. Isn't this basically metalearning?

Probabilistic weight updates? That is pretty much it.

"In addressing each of these issues we needed to make changes to the model, something that would be impossible with a black box classification algorithm, but which is central to the model-based machine learning approach. With these model changes in place, the Clutter prediction system is now deployed as part of Office365, helping to remove clutter emails from peoples’ inboxes. Figure 4.27 shows a screenshot of the system in action, all using model-based machine learning!"

7:45pm. http://mbmlbook.com/Recommender.html

I'll stop here for the time being. I really had enough."