https://www.reddit.com/r/MachineLearning/comments/9yeh2s/d_can_logistic_regression_model_an_evolving_value/
Today I will finish the rest of the book. I asked this question on the ML sub. After I am done with this, I think I will study some of the papers I linked to yesterday and that course. After that I think I will just start programming.

https://www.youtube.com/watch?v=AfrU2wHQnrs
Blake Richards - Deep learning with segregated dendrites

Also this talk. Let me take a look at this talk first.

For the past few days I've been so asspained by the stabs in the dark. In comparison, doing the kind of work I am good at and scaling up is something that will be a welcome change. That having said, it is not in my nature to just give up.

It is in my nature to give up on the way I've been doing it and that is an important distinction.

Once I've scaled up and learned more, I might give another shot at making a breakthrough, but that won't be anytime soon, that is for sure.

Before I resume programming, what I will aim for is to get an understanding of how Bayesian RL is supposed to work.

I need some guidance and NNs are just not offering it. The way I made Spiral and the way I program is to understand what should be and move closer to it. I start with a hypothesis and chisel it out of reality. With small steps, I get closer to the light.

In contrast the ML work that I did just made me fall deeper and deeper into darkness the more I worked at it.