"https://arxiv.org/abs/1811.01458 Bayesian Action Decoder for Deep Multi-Agent Reinforcement Learning
Maybe I'll take a look at this too.

https://csc2541-f17.github.io/
CSC2541: Scalable and Flexible Models of Uncertainty

Oh, this could be interesting.

Oh, and I completely forgot about the course from yesterday. The webpage for it was down so I could not check it out.

https://faculty.ozyegin.edu.tr/melihkandemir/sml-course/
Statistical Machine Learning Course

It is finally back online.

https://medium.com/@peadarcoyle/why-would-i-ever-need-bayesian-statistics-4cf844c4a23a

http://mlg.eng.cam.ac.uk/yarin/blog_2248.html

2pm. Agh, there is a lot to go through. Let me read the BNN paper as planned.

...The paper is trash. Let me go for the dropout one.

2:05pm. "We
extend Variational Dropout to the case when
dropout rates are unbounded, propose a way to
reduce the variance of the gradient estimator
and report first experimental results with individual
dropout rates per weight. Interestingly, it
leads to extremely sparse solutions both in fullyconnected
and convolutional layers."

Oh, this is interesting. Did they crack prunning?

2:05pm. "Recently it was shown that dropout can be seen as a special
case of Bayesian regularization (Gal & Ghahramani, 2015;
Kingma et al., 2015). It is an important theoretical result
that justifies dropout and at the same time allows us to tune
individual dropout rates for each weight, neuron or layer in
a Bayesian way.
Instead of injecting noise we can regularize a model by
reducing the number of its parameters."

This does look extremely promising. I'll go with this if it does what I want and is not too complicated to implement.

2:15pm. I am looking up citations for the VD paper.

https://scholar.google.hr/scholar?as_ylo=2018&hl=hr&as_sdt=0,5&sciodt=0,5&cites=11014728550012194230&scipsc=

https://arxiv.org/abs/1803.10123
"We suggest a novel approach for the estimation of the posterior distribution of the weights of a neural network, using an online version of the variational Bayes method. Having a confidence measure of the weights allows to combat several shortcomings of neural networks, such as their parameter redundancy, and their notorious vulnerability to the change of input distribution ("catastrophic forgetting"). Specifically, We show that this approach helps alleviate the catastrophic forgetting phenomenon - even without the knowledge of when the tasks are been switched. Furthermore, it improves the robustness of the network to weight pruning - even without re-training."

This is a goldmine. It seems this is what I am going to be studying for the next month - everything Bayesian.

2:20pm. https://arxiv.org/abs/1810.03958
Fixing Variational Bayes - Deterministic Variational Inference for Bayesian Neural Networks
"Bayesian neural networks (BNNs) hold great promise as a flexible and principled solution to deal with uncertainty when learning from finite data. Among approaches to realize probabilistic inference in deep neural networks, variational Bayes (VB) is theoretically grounded, generally applicable, and computationally efficient. With wide recognition of potential advantages, why is it that variational Bayes has seen very limited practical use for BNNs in real applications? We argue that variational inference in neural networks is fragile: successful implementations require careful initialization and tuning of prior variances, as well as controlling the variance of Monte Carlo gradient estimates. We fix VB and turn it into a robust inference tool for Bayesian neural networks. We achieve this with two innovations: first, we introduce a novel deterministic method to approximate moments in neural networks, eliminating gradient variance; second, we introduce a hierarchical prior for parameters and a novel empirical Bayes procedure for automatically selecting prior variances. Combining these two innovations, the resulting method is highly efficient and robust. On the application of heteroscedastic regression we demonstrate strong predictive performance over alternative approaches."

They 'fix' variational Bayes.

This paper only came out a month ago.

I can't deal with this all at once. Let me finish the dropout paper first.

2:25pm. "It is beneficial
to use continuous noise instead of discrete one because
multiplying the inputs by a Gaussian noise is equivalent
to putting Gaussian noise on the weights."

Actually, the thought of using Gaussian noise to multiply the output never even occurred to me.

"Gaussian Dropout training is equivalent to stochastic optimization
of the expected log likelihood (2) in the case
when we use the reparameterization trick and draw a single
sample W ∼ q(W | θ, α) per minibatch to estimate the expectation."
"