Let me start. Yesterday I was just too tired after studying up to 8pm to make a commit, so let me catch up now.
I found something big yesterday.

https://arxiv.org/abs/1811.03679
Practical Bayesian Learning of Neural Networks via Adaptive Subgradient Methods

It is this paper. It is what I am going to use to make my nets Bayesian. It is one of the missing pieces and will make learning significantly more roboust in RL. It will be great for exploration too. Poker is just bandits plus some pattern matching so I won't need curiousity for it. In fact, I do not want curiosity for poker as it would bias it by making it risk seeking. I am amazed that Bayesian NNs are also good against catastrophic forgetting.

Using the Hessian matrix to sample weight distributions from is a trully brilliant idea and it will work that much better with KFAC which is a block diagonal rather than the diagonal approximation used in the paper.

Now that I think about it, the FIM was always supposed to be a measure of uncertainty so it makes sense that it could be used to sample weights from. I am embarassed that I did not think of this myself.

That having said, even though I am going to go with this - I certainly need it to give me that leg up, this does not change my newfound feelings on Bayesian learning. Bayesian learning is to deep learning what functional programming is to imperative programming.

With this I can finally imagine the agents defeating the microstakes soundly. I was really afraid of doing it previously, but with this it will work. Using KFAC for both noise injection and stabilization will bring me trully far. I can trully see myself winning now.

It is amazing that pruning of all things is the major application of Bayesian learning. I had not thought it would turn out to be like this. I thought it would be a huge problem with no solution in sight, but here it is. The VD paper pretty much described how it should be done - neurons that have low signal to noise are the ones who should be trimmed.

With this piece in my hands, I will be able to dedicate myself to pure coding for the foreseable future. I'll race straight ahead until I reach the finish line.

That will come after I get through the backlog. Even excluding those two courses, yesterday I've built up a huge pile of papers that I need to go through and finishing all of this will take me a few more days, but after that I'll be set.

I am actually pretty excited to start my studies for the day, so let me do it. I'll start with the 92 paper by MacKay and then move on to the Vadam one.