I am up early. I can't sleep because of Hebbian learning.
https://papers.nips.cc/paper/3391-hebbian-learning-of-bayes-optimal-decisions.pdf
Hebbian Learning of Bayes Optimal Decisions

Let me finish the first slide from that course and then I'll take a look at this. I've been really reflecting on my affection for Hebbian learning for the past few days. I think I lost sight of what is the goal of it - having a good memory. I might be able to find something better than it, at least for implementation on GPUs. I need to decide what is really important. I need to review my decisions.

https://faculty.ozyegin.edu.tr/melihkandemir/files/2017/04/sml_lecture1.pdf

Right now I am finally on the course slides. The plan is to finish the two course slides, then move on to the Bayesian RL references from the Berkeley course and then dive into the Approximate Inference thesis. There are a bunch of other theses as well I want to dive into like the one from Yarvin Gal and the huge Expectation Propagation stuff. I think I might be able to skim most of it today though. Yesterday I spent the full day just reading and today will be much like that.

Let me get on with it.