https://www.reddit.com/r/reinforcementlearning/comments/9z2b0t/r_practical_bayesian_learning_of_neural_networks/
/u/tihokan: "Are you planning to use it for exploration, similar to Noisy Networks for Exploration but with a more principled noise distribution?"

/u/abstractcontrol (me):

```
Yeah. I've been inspired by Wolfgang Maass's [talks](https://www.youtube.com/watch?v=zpDzSonhymE&index=7&list=PLgKuh-lKre10qVKXL6EqR08qxyHf8R7-A) that the brain essentially creates noise for some reason. I already know that combining noise and greedy algorithms makes powerful optimizers for constraint satisfaction tasks, and with NNs in particular dropout always struck me as not being good enough.

The same goes for the paper you've linked me to - having the noise coefficients optimized directly by SGD cannot be right as the incentive is there for it to push them straight to zero. With that paper, there is no reason at all to think that the algorithm will do what you want which is push the network to explore in the direction of most uncertainty.

From the perspective of exploration, I think this method will work quite well on its own, but I had an idea how I could add intrinsic motivation to the mix.

Assuming `V` and `V'` are values the present and next timestep, then for a net with two value heads it is possible to define a confusion function as `confusion V = abs (fst V - snd V)`. The intrinsic reward could then be calculated as `reward_intrinsic = confusion V' - confusion V`. With [multiple value heads](https://arxiv.org/abs/1602.04621) the confusion function would be `confusion V = sum (V - min V) / (length V - 1)`. The logic of having a model that is capable of estimating uncertainty is what will make this scheme possible.

Apart from exploration, Bayesian NNs have other good properties. [Resistance to catastrophic forgetting](https://arxiv.org/abs/1803.10123), [resistance to adversarial examples](https://arxiv.org/abs/1806.00667), [resistance to overfitting](https://arxiv.org/abs/1701.05369). In that last paper, I found it interesting how with sufficient regularization the network could stop memorizing random labels, this is something that is not the case with ordinary dropout.

I know that standard dropout has been shown to not work well for RL. In the bootstrap paper it is shown that it does not produce good uncertainty estimates, but there are different kinds of noise and it might turn out that we just need to try harder. Good uncertainty estimates should be a massive advantage.

I'll replace the particular scheme in the Badam paper with a KFAC variant. A few days ago the thought that I might be able to stabilize Bayesian NNs with KFAC is what started the search for this method, and sampling from the weight distribution produced by it can only be beneficial.

The only risk is using the Hessian to estimate uncertainties might be insufficiently Bayesian - the paper does not have much in the way of tests, but the method is just so simple that I can be confident in it working purely from that. In probabilistic models the [FIM](https://wiseodd.github.io/techblog/2018/03/11/fisher-information/) is literally supposed to be an uncertainty estimate of the curvature, so it would be surprising if there was no way to adapt it as a part of estimating the weight distribution.

That about covers my view in favor of this method at this time.

The only disadvantage of this method is that there is now a need to tune the noise level in addition to the learning rate. And that despite everything it might not work, but that argument will be resolved by testing.
```