I think my study time is over. Tomorrow, I am going to give myself some time off. Today even though I was supposed to be reading the thesis, I could barely focus as I was thinking about plastic RNNs the whole time.
Basically between the instability of modulation and having to implement reversible learning to make it at all scalable, I do not think I will be going for it for the time being. I'll leave it for 2020 and later.

It is just too hard right now. I need an extra push to make it work.

I am trying to decide whether temporal connections and such would be a good substitute, and the answer is to an extent. Unfortunately, if you want input reconstruction I am convinced that there is no better mechanism than plasticity. It is like having an autoencoder right there in the layer.

It is going to be necessary at some point, but I do not think I will need it for poker. I am not exactly sure what it would be the most necessary for though.

Maybe understanding exactly which sort of task it trully needs is something I should figure out before I grasp for it again.

Let me make that a rule.

I've been obsessed with the plastic RNN for the whole day today and I need to give it a rest. I've been way too early with differentiable plasticity.

I think tomorrow it will be time to start preparing for the next step. Hopefully Bayesian RNNs won't fall flat on their face like my previous ideas.

I am really hyped about the new weight sampling scheme and investigating it will be my sole focus for the duration of December.

The last few days were amazing. It was like having a reaper's scythe at my neck and all I did was study and study like a crazed person. That is true despair.

I must admit, this recent development caught me completely off guard. I never thought something like this would come out. I thought memory would be solved before exploration and generalization, but now the gradient is pointing in a completely different direction.

Well, memory in particular came after everything else in nature, so maybe ML is following the same trajectory.

Since I do not feel like resting how about I read those expectation propagation papers for a bit and then stop after that?