"``` inl unzip tns =     inl {basic dim=dim'} = tns.unwrap     inl {bodies dim} = basic.unwrap     Struct.map (inl bodies -> facade {basic=Tensor.facade {bodies dim}; dim=dim'}) bodies ```
I need to think about the design of this a little.

8:45am. Ok, I'll do it directly.

```
inl unzip' f tns =
    inl {basic dim=dim'} = tns.unwrap
    inl {bodies dim} = basic.unwrap
    Struct.map (inl bodies -> facade {basic=Tensor.facade {bodies dim}; dim=dim'}) (f bodies)

inl unzip = unzip' id
```

I think this should do it. Let me implement `primal` and `adjoint`. As a side note, does `ref` work on the Cuda side?

8:50am.

```
    inl primals = View.unzip' (Struct.map (function {primal} | primal -> primal))
    inl adjoints = View.unzip' (Struct.map (function {adjoint} -> adjoint | _ -> ()))
```

Let me do it like this.

```
    inl on_non_nil ret B =
        match B with
        | () -> ()
        | B -> ret B
```

This along with a bunch of other stuff is going right out.

8:55am.

```
    inl dr s primal = {primal adjoint=s.CudaTensor.zero_like primal; block=()}
    inl drv s primal = {primal adjoint=s.CudaTensor.zero_like_view primal; block=()}
```

I'll make the second the default and add a zip here.

...Hmmm...I do not have a view zip at the moment.

Let me stop here for a bit. I said I would be taking it easy."