"2:05pm. The MER paper was pretty interesting.
https://brage.bibsys.no/xmlui/handle/11250/2571885
"...But through extensive exploration we show that it's far from trivial to apply MAML to object detection on natural images. However, we are able to use a simpler method inspired by MAML, Reptile. We show that a model pretrained using Reptile can be fine-tuned considerably faster than a model pretrained normally on object detection, but surprisingly it does not enable using fewer examples. In addition, we show that Reptile is able to speed up the development of deep learning models in practice. This is done by building a proof of concept tool and use this to test some example use cases."

2:20pm. Wow, this is 132 pages?

2:25pm. https://arxiv.org/abs/1807.05960
Meta-Learning with Latent Embedding Optimization

I am in the mood for these kinds of papers. Ah screw programming. Let me go with this for now."