Here is the full post that I've edited numerous times on Reddit. I think I have enough evidence to decide to ignore Reptile going forward. The model based meta-RL paper is over half a year old now and given that there haven't been any new results combining it with RNN, I can be confident in that prediction, though half a year is not overly long. But it is decently long.
"I am wondering about a few things:

1) Could Reptile be somehow combined with RNNs? So far all the work on MAML that I've seen has been on feedforward models.

2) Moreover, does it even make sense to want to combine them?

3) Even if that could be done, how could Reptile be combined with KFAC and similar optimizers such as Adam which use some approximation of the Hessian?

4) How exactly does this relate to Bayesian NNs which are also resistant to catastrophic forgetting?

**Edit**: To answer my own first two questions, I'll refer to the recent [lecture](https://youtu.be/lXJeYCRtFvI?list=PLkFD6_40KJIxJMR-j5A1mkxK26gh_qg37&t=3178) by Finn who recently published her [PhD thesis](https://www2.eecs.berkeley.edu/Pubs/TechRpts/2018/EECS-2018-105.html). In the part I've linked to, she notes that RNNs are more powerful than MAML and that in the case of policy gradients with zero rewards they can capture a lot of information about the dynamics unlike MAML.

This answer is actually somewhat orthogonal to the perspective which led me to ask the questions, but as the purpose of MAML is to avoid optimization difficulties surrounding RNNs, I'd guess it would defeat the point.

This perspective is not entirely consistent as in the [model based meta-RL paper](https://arxiv.org/abs/1803.11347) she covers later, it is explicitly noted that combining both approaches might get the benefits of both (as the RNN is poor at extrapolating to out of distribution samples.)

Based on the information from the Reptile CNN paper in the previous post, I'd bet against MAML improving consistency of RNNs. The authors of it had to switch to Reptile in order to get convnets to train due to optimization difficulties. In addition to that, they said Reptile did not improve sample efficiency. Hence the conclusion would be that the model based meta-RL paper is getting better results with MAML in places due to the RNN being hard to train and only that."

Let me do some programming now and then I'll stop for the day. Hopefully I'll get into the swing of things tomorrow.