The fatigue is really hitting me right now, so I'll call it a day. I've done a bit of work on the library to make everything go smoother, but it was not much. I was not kidding when I'd said that I'd take it easy. Hopefully I'll get more into it tomorrow.
I think now that I've made up my mind to ignore Reptile, I'll go straight towards my vision. I have that sort of contrary mind - if I am 100% sure of something, that is when I'll start looking for alternatives and things to do after it is done. I think this is a good trait to have as a programmer.