"4:45pm. Let me make `assert_broadcastable` a template.
4:50pm. Done, that should be a better choice than a complete redesign. I am not sure I could do a proper job of it even if I tried it.

```
assert (dim = out.dim') "The input and the output must have the same dimensions."
```

What is dim' here?

```
dim' = inl {dim} -> Struct.map (inl {from near_to} -> near_to - from) dim
```

Ah, so I had this shit. But what about the standard thing?

Seriously, could I be any less careful?

Let me remove that thing. It has to be a mistake in `segmented_init` that I am using it.

4:55pm. I am going to remove this branch from `index_example'`.

```
            | {from near_to} -> next (dyn 0 :: l)
```

```
assert (dim = out.dim) "The input and the output must have the same dimensions."
```

Also I will make it so that that diming a view returns its true size.

I am going to have to make a redesign.

I think I am going to have to remove the ranges from views after all. They are just not worth the extra difficulty them make to compare tensor dimesions. Tree views are extremely useful, but ranges are just dead weight.

Right now even if I let dim return the true dimensions, the comparison would break down in the presence of ranges.

AAAggggghhhh...

I can't deal with this right now.

What I will do tomorrow is make wrapper that just supports ranges over the basic tensor, but no tree views. That would be for the best.

Agh, I messed this up. But it is fine, I'll do it properly this time.

5:05pm. Let me stop here for the day.

Right now the integrity of the ML library is surely an absolute wreck. I can't even begin to guess at what breakages will be. I am not sure I can even track the changes that I did mentally.

But I guess that for a redesign like this, it makes sense to start with the tensors and work my way to the kernel.

Tomorrow, I will take care of the tensors and deal with the type inference thing in `init_seq`."