I'll admit, after programming for a full day, I am finally starting to get into it. The heat of struggle is starting to warm my soul. When I get through all of this and implement weight sampling it will ignite.
Arnold Salas and Stefan Zohren. I'll be sure to remember their names for giving me a gift such as this.

The Badam paper came out literally at the exact time that I most needed it. I'd like have missed weight sampling via the curvature on my own, but now that it has been pointed out to me I sure as hell will not miss it.

All the prerequisites that I need to make a good player have been met.