Let me start work for the day. I can easily do the type catch and type raise in an hour or two.
After that I get to move on to the ML library.

Really, the work seems like it will never end, but the first phase will finish once I get through this as I'll finally have enough to make a good player.

Just finding the right algorithm was the main difficulty of 2018 - having it in my hand means I can stop wandering around and start digging it out.

I'll crush the poker donks and move on to more important things after that - there will be plenty of challenges left before a complete mastery is attained and the Singularity era starts.

But the most important thing is that I finally establish a handhold and arrest my fall.