"10:50am. Done. Now the type inference for `init_seq` is actually simpler than for the other kernels.
This solution for type inference is hacky, but also highly effective. I am quite pleased with it as it is definitely a good solution. Expections are slow in .NET, but I am not going to be using them often. I think it should be fine.

10:55am. Right now I am not exactly sure what I should do next. Where was I last time? Let me determine this and I will take a break.

I think it had something to do with `assert_broadcastable`..."