"``` {bodies=<function>; dim=[lit 128i64, lit 784i64]} {basic=<function>; tree_dim=[[], []]} ```
I forgot to wrap a view in the test.

4:05pm. Stop reading social media. I need to continue.

```
Error trace on line: 617, column: 24 in file "Learning".
            inm data = segmented_init {dim=input.span_outer,outer} {bias=const one; input=load input}
                       ^
Error trace on line: 95, column: 7 in file "View".
    | .(_) & x ->
      ^
Error trace on line: 96, column: 9 in file "View".
        if module_has_member x data then data x
        ^
Error trace on line: 97, column: 14 in file "View".
        else methods x data
             ^
Cannot find a member named span_outer inside the module.
```

Did the wrap, now I need to get rid of span_outer.

4:10pm.

```
Error trace on line: 168, column: 76 in file "Tensor".
                            {(rest from) with bodies=Struct.map2 (inl size (!unconst ar) -> const {ar with size=size :: self}) size self; dim=near_to - from :: self}
                                                                           ^
Error trace on line: 325, column: 13 in file "Core".
inl unconst x = x()
            ^
Expected a function in function application.
```

That `map2` might have to be `map2'`. Let me put in a print. Today is turning out just as horrible as I expected it would be. Will I even get this to run in the next hour or so?"