"Let me put in some print statements. To be honest, I had quite enough of this. I knew that the breaks would be bad, but this is really, really bad. I was sure I would finish Mnist today, but now I am thinking of simply leaving this for tomorrow.
Fixing type errors for over 4.5 hours straight is taking it toll on me. This kind of work is the very definition of drudgery. I sure wish Spiral was Haskell in times like these.

Well...let me at least take a look in what is going on in that tensor. Why is it not `const`'d?

```
{bodies={adjoint=<function>; primal=<function>}; dim=[lit 128i64, lit 785i64]}
{bodies={adjoint=<function>; primal=<function>}; dim=[lit 128i64, lit 785i64]}
```

This is missing a block, though that does not matter.

```
{bodies={adjoint=<function>; block=[]; primal=<function>}; dim=[]}
{adjoint=var (float32); block=[]; primal=lit 1.000000f32}
{bodies={adjoint=<function>; block=[]; primal=<function>}; dim=[]}
var (float32)
```

Right. The input here does not have an adjoint....

5:05pm. Aggggghhhh...I am going to have to infer the max of types.

```
                    |> Struct.map (function
                        | {primal adjoint} as x -> {x with adjoint=adjoint 0}
                        | x -> x
                        )
```

Let me take this out.

5:10pm.

```
inl elem_type = type init (index_example' dim) // TODO: Infer the union of types here.
```

I give up. I cannot do this today. The redesign is proving to be a lot more complicated than I thought it would be.

...No, I won't even take it out. I'll just leave it for tomorrow.

5:15pm. Let make the last commit for the day. To be honest, I am not even sure how I should resolve this issue.

I guess I'll make some kind of unification scheme. I'll use the `Union.infer` to get all the types, and then unify them.

```
inl index_example' dim =
    Liple.foldr (inl x next l ->
        inl rec loop = function
            | {} as x ->
                inl {k x} = module_foldr (inl k x _ -> {k x}) x ()
                {$k = loop x}
            | _ -> next (dyn 0 :: l)
        loop x
    ) dim (Tuple.rev >> Tuple.unwrap) ()
```

Let me get rid of this.

5:25pm. This will take some engineering work. I am still thinking about. Well, plenty of unexpected things pop up while one is doing major redesign, so delays should not be too surprising."