"8am. https://arxiv.org/abs/1811.11214 Understanding the impact of entropy in policy learning
8:05am. Let me start the day with a commit. I had some time to think about exactly I wanted to do last night.

I think I am going to bring back `concat`. Before I do automatic inference, it might be good to start out with something lighter just so I can get Mnist to work. Having `concat` will allow me to make the unification function that I'll be able to reuse later in automatic inference. I'll let `segmented_init` take the `elem_type` as argument for the time being. I can't deal with this all at once. Once I have a grip on the library, then I'll be able to move. Complexity is something that can only be added gradually.

But before that, let me chill. Damn, I am sick of work already and there is at least ten hard days of this. I consider this redesign work in the worst sense of the word - it will be such a relief when I am done with it.

The best thing is to just do it one at a time. There is no rushing through this.

For now let me read that paper, chill a bit and then I will start."