Just wasting time is simply the best, but now I have to do some work. Let me do programming for a while.
I must surmount this challenge. Let me start by dealing with segmented_init.

Damn, if my hopes regarding weight sampling come true, the future is going to be so great. But if not, I am not sure what I am going to do.

Just like I abandoned trading, I might need to do the same with deep RL and look for my opportunity somewhere else until the field gets there. I won't even think about this possibility for the next two weeks.

If the weight sampling experiment fails, I am not going to have any choice but to accept it.

My strategy for the past 3.5 years was perfect - I did everything to the utmost of my ability and made it so that I take advantage of all the luck I can get. But the luck is still a significant factor in all of this. And I do not have infinite will.

The curvature-based weight sampling is my last and greatest hope. It is time to face the nightmare.