```     inl rec unify_dual =         Struct.map2 (inl s r ->             match s,r with             | {primal adjoint}, {primal adjoint} ->                 assert (eq_type s r) "The two structures must be the same type."                 s             | {primal adjoint}, _ ->                 assert (eq_type primal r) "The two structures must be the same type."                 s             | _, {primal adjoint} ->                 assert (eq_type s primal) "The two structures must be the same type."                 r             | _, _ ->                 assert (eq_type s r) "The two structures must be the same type."                 s             ) ```
This should cover most of my use cases. Note that if `s < r` then this unification won't be right, but a type error will occur in the kernel instead so it is fine.

Let me make the two upscale functions. The scalar ones should be trivial, but I am honestly not sure about...

Ah, no wait, it is not what I am thinking is it. I do not need to give them an adjoint, it would be fine if I make them `{primal block=()}`.

Ok, so...

```
    inl upscale_scalar elem_type x =
        match elem_type with
        | {primal adjoint} -> const {primal=x; block=()}
        | _ -> const x
```

Yeah, this should be right. I can pass this into `segmented_init`. The tensor will be more difficult, but thankfully I'll just use `update_body'`. Good thing I made that. Just how much of a bother is making this?

Let me finish this and I'll consider taking a break.