Why is this happening now?
```
store=inl {from=(!zip from) to=(!zip to)} -> inner_loop {body=inl {item i} -> to i .set (from item .get)}
                      ^
```

So this is where it happening.

```
[lit 1i64]
[]
```

The dimensions of the adjoint and primal are different.

```
            s.CudaFun.init_seq {dim} <| inl b k ->
                init b k .out
                |> Struct.map (function
                    | {primal adjoint} as x -> {x with adjoint=adjoint 0}
                    | x -> x
                    )
```

Ah, it is because I put this stupid thing in without thinking about it. I need to remove this.

```
        {
        out=View.wrap dim out
```

Let me also do the wrap here.

Oh, it compiled. Great. Let me give it a spin.

```
Training: 0.1836399253414
Testing: 0.113177325552664(9620/9984)
Training: 0.0643514293120601
Testing: 0.095317505673171(9671/9984)
Training: 0.0305183720871961
Testing: 0.0884926123858(9712/9984)
Training: 0.0138830112601316
Testing: 0.0851460069254524(9749/9984)
Training: 0.00584066336190688
Testing: 0.0820976331742713(9754/9984)
```

Yeah, it works now. Great.

Wonderful. I am guessing this would be a good time to plug in those symmetric multiplies into `kfac`.

Let me put the timer into Mnist and then I will do that. It won't be hard. It will be my last move for the day.